Updated SideBar UI + improved code structure to make it more modular
