+++ apps/web/src/lib/components/Launch.svelte
@@ -12,6 +12,7 @@
     import { ProjectType } from "../utils/types";
     import { Button } from "./ui/button";
     import { SunMedium, Moon } from "@lucide/svelte";
+    import ScreenSizeFallback from "./generic/ScreenSizeFallback.svelte";
 
     /**
      * Sets Project Type
@@ -75,11 +76,5 @@
 </main>
 
 
-<!-- FallBack in case screen size is too small -->
-<main
-    id="body"
-    class="w-screen h-screen overflow-hidden flex justify-center items-center md:hidden">
-    <p class="text-3xl font-bold text-center font-geist">
-        Screen Size too Small
-    </p>
-</main>
+<!-- Displayed if screen size is too small -->
+<ScreenSizeFallback />
