+++ apps/web/src/lib/components/generic/ScreenSizeFallback.svelte
@@ -1,0 +1,8 @@
+<!-- FallBack in case screen size is too small -->
+<main
+    id="body"
+    class="w-screen h-screen overflow-hidden flex justify-center items-center md:hidden">
+    <p class="text-3xl font-bold text-center font-geist">
+        Screen Size too Small
+    </p>
+</main>
